TradeHistoryParams fix errors

+++ xchange-core/src/main/java/com/xeiam/xchange/service/polling/trade/TradeHistoryParamsAll.java
@@ -6,7 +6,7 @@ import com.xeiam.xchange.currency.CurrencyPair;
  * Generic {@link TradeHistoryParams} implementation that implements all the interfaces in the hierarchy and can be safely (without getting exceptions, if that all the required fields are non-null) passed to any implementation of {@link com.xeiam.xchange.service.polling.PollingTradeService#getTradeHistory(TradeHistoryParams)}.
  */
 public class TradeHistoryParamsAll implements TradeHistoryParamsTimeSpan, TradeHistoryParamCount, TradeHistoryParamsIdSpan, TradeHistoryParamOffset, TradeHistoryParamCurrencyPair {
-  private Long count;
+  private Integer count;
   private String startId;
   private String endId;
   private Long endTime;
@@ -15,13 +15,13 @@ public class TradeHistoryParamsAll implements TradeHistoryParamsTimeSpan, TradeH
   private CurrencyPair pair;
 
   @Override
-  public void setCount(Long count) {
+  public void setCount(Integer count) {
 
     this.count = count;
   }
 
   @Override
-  public Long getCount() {
+  public Integer getCount() {
 
     return count;
   }
